+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/config/IOperationAbilityConfig.kt
@@ -1,7 +1,5 @@
 package site.siredvin.peripheralium.api.config
 
 interface IOperationAbilityConfig {
-    val isInitialCooldownEnabled: Boolean
-    val initialCooldownSensetiveLevel: Int
     val cooldownTresholdLevel: Int
 }